requirements.txt
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==2.3.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
mailtrap==2.0.1
MarkupSafe==2.1.3
Pillow==10.1.0
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
urllib3==2.1.0
Werkzeug==2.3.3